libshvrpc: use intmax_t for client IDs instead of long long
The actual maximum we are able to handle is intmax_t and not long long.
Thus client IDs list should work with this type to be consistent with
libshvcp.

+++ include/shv/rpcmsg.h
@@ -158,7 +158,7 @@ struct rpcmsg_meta {
 	 * Valid for :c:enumerator:`RPCMSG_T_REQUEST`,
 	 * :c:enumerator:`RPCMSG_T_RESPONSE`, and :c:enumerator:`RPCMSG_T_ERROR`.
 	 */
-	long long *cids;
+	intmax_t *cids;
 	/** Length of the ``cids`` array. */
 	size_t cids_cnt;
